FI-006_Inventory2.sql
Gets fields from get_fi006, get_sa001 to be used in the Dashboard Inventory.
The data is grouped by month so as to get the Value_in_EUR per month. 
We also get the next month Value_in_EUR as well as the month after that.
We then sum them together and devide them by the total number of days so as to get the Days_of_Stock.