KPI.sql
The 2nd request used in the Dashboard 'Weekly KPI dashboard'. The 1st one I did not write.
Gets fields from kpi."Def_Serie" and "Dat_Serie_Tableau" to be used in the Dashboard 'Weekly KPI dashboard'.
Gets the fields that are seized per site (can be found in kpi."Def_Serie" which Site seizes which fields).
Gets the values where the date is < to Sunday (example we are Monday 26/06/2017, the Sunday is 25/06/2017). The values where the date is = to Sunday and the values where the date is equal to the Past Sunday (example we are Monday 26/06/2017, the Sunday is 18/06/2017).